requirements.txt
# Core Flask
Flask==3.1.2
Werkzeug==3.1.5

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
PyMySQL==1.1.2
SQLAlchemy==2.0.46
alembic==1.13.1

# Authentication
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1

# Data Processing
pandas>=2.2.2
numpy>=1.26.0,<2.1
openpyxl>=3.0.0

# Machine Learning
scikit-learn==1.6.1
joblib

# AI Integration
openai>=1.0.0
requests>=2.31.0

# Environment & Configuration
python-dotenv==0.19.0

# Production Server
gunicorn

# Utilities
cryptography

# Payment Processing
stripe==7.9.0

# Web Scraping & HTTP
beautifulsoup4==4.12.3
selenium==4.16.0
lxml==5.1.0

# PDF Generation
reportlab==4.0.9
PyPDF2==3.0.1

# Barcode/QR Processing
python-barcode==0.15.1
qrcode==7.4.2
Pillow==10.2.0
pyzbar==0.1.9
